requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
